create a makefile to build the project
