--- src/Operator.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Darshil Patel on 4/18/24.
-//
-
-#include "Operator.hpp"
