--- src/Operator.hpp
@@ -1,9 +1,0 @@
-#ifndef OPERATOR_HPP
-#define OPERATOR_HPP
-
-// You might want to define specific operator functionalities or enumerations here
-class Operator {
-    // Placeholder for operator specific properties and methods
-};
-
-#endif // OPERATOR_HPP
