--- truthy src/CLIInterface.hpp
@@ -1,9 +1,0 @@
-#ifndef CLI_INTERFACE_HPP
-#define CLI_INTERFACE_HPP
-
-class CLIInterface {
-public:
-    void run();
-};
-
-#endif // CLI_INTERFACE_HPP
