--- truthy src/ExpressionEvaluator.hpp
@@ -1,11 +1,0 @@
-#ifndef EXPRESSION_EVALUATOR_HPP
-#define EXPRESSION_EVALUATOR_HPP
-
-#include <string>
-
-class ExpressionEvaluator {
-public:
-    bool evaluate(const std::string& expression);
-};
-
-#endif // EXPRESSION_EVALUATOR_HPP
