--- truthy src/main.cpp
@@ -1,7 +1,0 @@
-#include "CLIInterface.hpp"
-
-int main() {
-    CLIInterface cli;
-    cli.run();
-    return 0;
-}
